Running AES encryption on new thread
