+++ flutter-app/vault/lib/Classes.dart
@@ -18,5 +18,7 @@ class Vault {
 
   VaultState state = VaultState.Locked;
   String codename;
+  String name;
+
   String accessToken;
 }
